// web/src/shared/api/generated/model/getAuditLogsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type GetAuditLogsParams = {
actor?: string;
resourceType?: string;
from?: string;
to?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};