// web/src/shared/api/generated/model/pageDataOfAuditLogResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AuditLogResult } from './auditLogResult';

export interface PageDataOfAuditLogResult {
  items?: AuditLogResult[];
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pageNumber?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pageSize?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalItems?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalPages?: number | string;
}